Extract mock_db
