+++ fuel-txpool/src/lib.rs
@@ -4,6 +4,9 @@ pub mod service;
 pub mod txpool;
 pub mod types;
 
+#[cfg(test)]
+mod mock_db;
+
 pub use config::Config;
 pub use fuel_core_interfaces::txpool::Error;
 pub use service::{Service, ServiceBuilder};
